Refactor distinct dates.
